Fix slow licence versions 4 charge workflow query (#1781)
https://eaflood.atlassian.net/browse/WATER-3745

This issue was found whilst working on [Remove node-cron for BullMQ jobs](https://github.com/DEFRA/water-abstraction-service/pull/1780). In the module `charge-versions` node-cron is used to kick off a process that creates a job for every licence version ID not in the charge workflow.

We needed to recreate that with a BullMQ repeatable job so got cracking. The current process is scheduled to kick off every 6 hours but we wanted to see it running so amended the schedule to every minute.

Well, the job triggered and then we waited for the `onComplete` to fire. And waited, and waited, _and waited_! 🤔

It just didn't seem to be completing. So, we traced through the code and ended up at `src/lib/connectors/repos/queries/licence-versions.js`.

We ran it locally, just to see if it returned anything but instead found ourselves waiting for 18 minutes for it to complete! 🤯😱

Picking a different tool and switching off the app, on repeated runs we got timings of 6 to 7 minutes. It didn't seem like an overly complex query but it was certainly 'messy'. Seeing `SELECT *` all over the place was an immediate 'smell' that the proper care hadn't been taken with it.

We're no SQL experts, but a quick Google on performance tips highlighted `NOT IN` should be avoided and be replaced with `LEFT JOIN .. IS NULL`. When we tidied up the query and dropped the `NOT IN`, the timings went down to less than a second. ✊😎