Convert licence not in charge version from cron
This is the first of the jobs we've encountered that was using node-cron to run the part of the process that determined which licence versions were not in the charge version workflow. For each one found it was creating a job in the `src/modules/charge-versions/jobs/create-charge-version-workflows.js` queue.

Initially, we thought it was just the usual coding inconsistency. Why was this a node-cron job when others were using BullMQ repeatable jobs? But then we realised that the underlying query in its original form was taking 5-6 minutes on a MacBook Pro. So, goodness knows how long it was taking on the production servers. By default, workers need to have completed jobs in less than 30 seconds else they become stalled.

With the work done in [Fix slow licence versions 4 charge workflow query](https://github.com/DEFRA/water-abstraction-service/pull/1781) we can now handle the first part as a repeatable 'job'. This means the whole process will be handled by BullMQ once this change is merged.