Convert check 4 updated invoice accounts from cron
This was more like what we were expecting. Useing node-cron to schedule the kick-off job was purely that. It was just adding an empty job to the queue unike the previous one which was running a query to populate the jobs with.

As part of the conversion we did also refactor both the job and its tests to bring it inline with the previous job we worked on.