Remove node-cron for BullMQ jobs (#1780)
https://eaflood.atlassian.net/browse/WATER-3745

This forms part of the work we are doing to [move all background processes to a background instance of the app](https://github.com/DEFRA/water-abstraction-service/pull/1766).

Another step to that is [Refactoring message-queue to centralise job registrations](https://github.com/DEFRA/water-abstraction-service/pull/1775). At the moment each module does its own thing when it comes to registering BullMQ jobs. The problem is it makes it really hard to see just what on earth is going on in the background of the service.

It has also led to inconsistencies, one of which is using both repeatable BullMQ jobs and [node-cron](https://github.com/node-cron/node-cron) to achieve the same thing.

This change is targeted where **node-cron** is being used to schedule creating jobs. This can easily be replaced by creating a [repeatble job](https://docs.bullmq.io/guide/jobs/repeatable) at startup that operates on the same cron schedule.

We'll also make the schedules configurable so we can have more control over when they run across the different environments.

** Notes

- Convert licence-not-in-charge-version-workflow from node-cron

This is the first of the jobs we've encountered that was using node-cron to run the part of the process that determined which licence versions were not in the charge version workflow. For each one found it was creating a job in the `src/modules/charge-versions/jobs/create-charge-version-workflows.js` queue.

Initially, we thought it was just the usual coding inconsistency. Why was this a node-cron job when others were using BullMQ repeatable jobs? But then we realised that the underlying query in its original form was taking 5-6 minutes on a MacBook Pro. So, goodness knows how long it was taking on the production servers. By default, workers need to have completed jobs in less than 30 seconds else they become stalled.

With the work done in [Fix slow licence versions 4 charge workflow query](https://github.com/DEFRA/water-abstraction-service/pull/1781) we could now convert the first part as a repeatable 'job'. This means BullMQ will handle the whole process.

- Convert billing.find-update-invoice-accounts from node-cron

This was more like what we were expecting. Using node-cron to schedule the kick-off job was purely that. It was just adding an empty job to the queue. This is unlike the previous one, which was running a query to populate the initial jobs.

As part of the conversion, we did also refactor both the job and its tests to bring it in line with the previous job we converted.