Add new StartUpJobsService
We've created a sinmgle place to register jobs with `JobRegistrationService`. We also updated the QueueManager to handle cleaning queues upon registration.

SO, at this time the only thing the module level register subscribers are doing is add the initial jobs. These are repeatable jobs we add during startup which then BillMQ manages the scheduling of.

Again, it makes sense to see what jobs we're kicking off during app startup in one place. So, we've created the `StartUpJobsService` to handle this process going forward.