Use fork mode for service-background
For a standard web-app or API you would typically run it using [cluster mode](https://pm2.keymetrics.io/docs/usage/cluster-mode/). This is some cleverness built into Node.js which allows multiple processes to be spawned that still share the same ports. How many is determined by the number of CPU's available. It's a codeless way to scale an app based on the resources available. To that pm2 will also act as a load balancer between the processes.

By default pm2 runs apps using fork mode (would love to link to a page but there ain't none!) This is the equivalent of what you do when running it locally. If you wanted to start a second instance you'd have to run it on a different port.

The reason for the change is because we are not using the background instance as a web app or API. Instead, we're using it to run processes in the background that should only be run once. What is happening in production at the moment is

- normal service starts up using pm2 with config set to cluster mode and `max` instances. Assume pm2 finds 2 CPU's so creates 2 instances of the app
- each app will instantiate BullMQ workers as part of start up (it will also try to register the queues in parallel!) Current count is we have 13 of them, so that's 26 additional processes started behind the scenes on top of our 2 apps
- a job is added to a queue and both workers created for the queue fight to process it
- then double all this because we have 2 EC2 instances running in AWS

A background job is only expected to be run once so typically you'll designate one of your EC2 instances to handle them. However, we treating our background service like a scaling web-app which makes controlling and debugging much more complicated.

We've still to think about how we handle a worker per EC2 instance (it may be fine). But within on an instance there should just be 1 background service running.