Only create worker and scheduler in background
We coming to the end of our changes to support splitting the work of the service project. Both foreground and background need to know about the queues. But only the background instance needs to create the instances of Worker and QueueScheduler to process the queues.

So, we add a new check (similar to what we did with the jobs that are using `fork()`) into the `QueueManager` to only create them if we are the background instance.