Replace supported source sync with migration
https://eaflood.atlassian.net/browse/WATER-3750

We currently populate the supported sources lookup table with the contents of a CSV file which is pulled down from S3 every 6 hours. Since the supported sources don't need to be updated anywhere near this often (if at all!) we instead remove all the code to do with this syncing, and replace it with a standard migration.