Remove unneeded dependencies (#1708)
https://github.com/DEFRA/water-abstraction-team/issues/26

Our `package.json` file has a number of dependencies hanging around which are no longer used. We identify them using `npx depcheck` and uninstall them accordingly.

Note that the following dependencies were identified as not being used, but they are used as far as we're aware, just not referenced in a way that depcheck recognises:
- `bookshelf-case-converter-plugin`
- `bookshelf-json-columns`
- `db-migrate-pg`

We make a couple of small code changes:
- One file requires `@hapi/joi` but in `package.json` we are including `joi` -- `@hapi/joi` only works as it's currently introduced as a dependency of a dependency. We therefore update the file to directly use `joi`.
- One file requires `request-promise` which again is only being introduced as a dependency of a dependency. Elsewhere in the code we use `request-promise-native`, which is included in `package.json`, so we update this file to follow suit.

Finally, we install some packages which are only currently included as dependencies of dependencies, to ensure that we are explicit about which packages our service uses.

* Remove unneeded dependencies

We run `npx depcheck` and remove the dependencies it doesn't find. We manually check to ensure each one isn't used; `bookshelf-case-converter-plugin` and `bookshelf-json-columns` are referenced in a way that depcheck doesn't recognise and therefore we keep them even though depcheck thinks they're unused.

* Remove unused dev dependency

Again, we run `npx depcheck` and remove the unneeded dev dependency it finds. Note that it claims `db-migrate-pg` isn't needed and we cannot find a reference to it, but the docs for `db-migrate` do state that it's needed. We may do further checks to see if this is the case but for now we leave it.

* Remove `mkdir`

Merging the current state of `develop` reintroduced the `mkdir` package, so we remove it again

* Fix `@hapi/joi` require statement

One of our files requires `@hapi/joi`, when we should just be requiring `joi`. This works because `@hapi/joi` is still a dependenchy of one of our dependencies, but this may change in future so we should directly require the `joi` package we have installed

* Fix `request-promise` dependency

One of our files requires `request-promise`, which is installed as a dependency of a dependency. Elsewhere we are using `request-promise-native`, so we change this file over to use it as well.

* Add missing dependences to `package.json`

Some of our dependencies aren't in `package.json`; they are installed as dependencies of dependencies. We add them in since we want to be explicit about our dependencies.