Amend purpose use validation from charge upload
https://eaflood.atlassian.net/browse/WATER-3749

Currently, the upload validates the `charge_element_purpose` field against the existing purpose uses on the licence. The matches what happens in the UI and primarily seems to be to allow it to identify and populate primary and secondary purpose information.

However, the charging team have identified a number of NALD records that have the incorrect purpose assigned to them. To correct it, they have to update NALD which then eventually updates the WRLS system It's a complex process in NALD which then needs to wait a day for the correction to appear in WRLS.

They are aiming to use the upload to correct these on-mass, which means the validation shouldn't be looking at the existing data (because it's wrong!)

This change updates the validation to only check the entry in the CSV against the lookup list of purpose uses in the DB. When we come to save the data we will only populate the purpose information in the charge element for the purpose use. We'll leave blank the primary and secondary information.