Use correct method, rename it, and fix exports
We accidently used the existing `getPurposeUses` method in `testPurpose` when we should have been using the new `getPurposes`, so we correct this. We then rename it to `getPurpose` to make its intent clear (ie. we are getting a single purpose from the db). Finally, we fix how we export our modules.