Change how we access the db
The code we based our db lookup on wasn't correct; it was simply creating a new instance of a model, rather than pulling anything from the db. As a starter for 10 we change the code to something that "looks" correct. Note that this is not yet tested in any way, as none of our helper methods are unit tested and we haven't yet tried a manual test. It also completely breaks our unit tests, as we need to stub out `getPurpose` to match how we treat the other helper methods.