Remove catbox-redis from service
https://github.com/DEFRA/water-abstraction-team/issues/46

Catbox-redis is only used in the service to cache the address search data. The reason given for it's use is "caching avoids API rate limit being exceeded on subsequent identical requests".

As we are using v1.1 of the EA Address Facade which was developed with rate limiting built in this cache is not required and can be removed. As this is the only place catbox-redis is used the package can be removed once the cache has been removed from the address lookup plugin.