Start on new unit tests
We are creating a whole new set of unit tests, so that instead of testing Notify (ie. sending it invalid data and checking what it returns) we stub it out using Nock and instead test that our controller correctly handles valid and error responses from Notify. So far we've written a first draft of the first few tests.