Small clean up for notifyLater/
Actually, this and the notify/ endpoints both hit the same controller. Plus, we think the scheduling part of notifyLater/ is broken (see https://eaflood.atlassian.net/browse/WATER-3758) and they are effectively the same endpoint.

So, we can amend the nock request to exactly match what we were doing in the previous test. We'll look at refactoring out the duplication once we've got all the endpoints covered.