https://eaflood.atlassian.net/browse/WATER-3752
As part of the annual billing process it is possible that a pro-rata calculation may be required where the licence has been e.g. revoked or transferred.

In these instances the ratio of (billable days/ authorised days) should be applied to the charge calculation. This doesn't appear to be happening.