Fix `libxmljs` unit tests
While trying to update `libxmljs` to the latest version, we found an issue where some of our unit tests start failing. This seems to be down to the newer version doing additional type checking. Before we update the version (which has other issues besides this!) we update the unit tests so we pass the correct type through.