Fix `libxmljs` unit tests (#1811)
* Fix `libxmljs` unit tests

While trying to update `libxmljs` to the latest version, we found an issue where some of our unit tests start failing. This seems to be down to the newer version doing additional type checking -- previously we could simply pass a parsed XML object to it and it would simply accept it (which is what the tests were doing) but now it throws an error if you try this.

Before we update the version (which has other issues besides this!) we update the unit tests so we correctly pass through a string and not parsed XML.

* Amend `mapper.test.js`

We update the mapper tests so that they correctly pass a string to `parseXml` instead of an already-parsed XML object.

* Remove commented-out code from `validator.test.js`

* Amend `xml-to-json-integration.test.js`

As with `mapper.test.js`, the test file `xml-to-json-integration.test.js` was reading a file and parsing it to XML, and then passing it to `mapXml`. We amend the tests to simply read it in the file and pass it to `mapXml` without parsing.