Remove `src/lib/response.js` (#1814)
* Remove `src/lib/response.js`

https://github.com/DEFRA/water-abstraction-team/issues/49

The file `src/lib/response.js` has an `envelope` method which wraps a response, however we only use it in a few places and don't make any use of the various features it has (converting keys to camel case, wrapping error responses, etc.). We therefore remove it completely and simply wrap each response ourselves.

* Delete `src/lib/response.js` and its test

* Replace use of `envelope` method

We use the `response#envelope` method in 3 files, so we replace these uses to simply wrap the return ourselves, ensuring we still return `error: null` to ensure compatibility with whichever parts of the UI calls them.