Remove docker files from repos (#1831)
DEFRA/water-abstraction-team#38

All our apps have a Dockerfile and .dockerignore. We even had some initial attempts at orchestration hidden away in GitLab and a fantastic effort was made with water-abstraction-orchestration.

But the files are very basic and appear to be slightly modified Hello, world versions rather than properly tailored multi-stage builds. Added to that the apps themselves are no-way ready for Docker.