Add initial SROC test billing data
This creates a new licence with associated SROC charge version, all dated for after 1 April 2022.

Left as it is, it would _not_ be selected for SROC supplementary billing. But the intent is to give us something we can use as a base to which we can then 'change' as per the expected scenarios. From this we can test, repeatedly, whether our algorithm is correct.