Add our own users fixture
Tl;DR; users created now match those either used for testing or referred to by the fixture data

Previously, we just used the `internal-users.yaml` fixture. But it creates a quite a number of users when we only need the one. Also, some of the CRM fixture data links to `acceptance-test.external@example.com`. But we weren't creating that user.

We'll probably not need it but for completeness, we now include that user in our test data.