Implement dynamic link to charge category
When we create a charge version we have to link it to a charge category. This becomes 'messy' for all the reasons set out in the comments we've added to `bookshelf-loader`.

Suffice to say it now works, and if we have to we can get more charge categories to use in this way in the future.