Add new SROC supplementary billing test data (#1868)
https://eaflood.atlassian.net/browse/WATER-3787

In order to test our new code to identify licences that need to be included in an SROC supplementary bill run we need some test data.

We want to ensure we have this in place rather than having to manual edit whatever licences happen to be at hand at the time we start to code.

So, this change adds a new test base for our SROC supplementary billing work.

**Notes

- Add initial SROC test billing data

This creates a new licence with associated SROC charge version, all dated after 1 April 2022. Left as it is, it would _not_ be selected for SROC supplementary billing. But the intent is to give us something we can use as a base to which we can then 'change' as per the expected scenarios. From this, we can test, repeatedly, whether our algorithm is correct.

- Add our own users fixture

Previously, we just used the `internal-users.yaml` fixture. But it creates quite a number of users when we only need one. Also, some of the CRM fixture data links to `acceptance-test.external@example.com`. But we weren't creating that user. We'll probably not need it but for completeness, we now include that user in our test data.

- Implement dynamic link to charge category

When we create a charge version we have to link it to a charge category. This becomes 'messy' for all the reasons set out in the comments we've added to `bookshelf-loader`. Suffice it to say it now works, and if we have to we can get more charge categories to use in this way in the future.