Add new presroc billing acc. change to fixture
https://eaflood.atlassian.net/browse/WATER-3787

We're adding to our new SROC fixture set a pres-roc licence that we will then update the billing account on.

In the UI this would mean updating both pre-sroc and sroc charge versions. In this scenario the 'licence' would be picked up for SROC supplementary billing because we added a new SROC charge version. But the new presroc charge version would be ignored.