Fix SROC licence
We spotted that the Ui was not showing the charge element section for the SROC charge version we created.

We realised that SROC charge versions must have a charge_purpose created as well. And to create the charge_prupose we would need some purposes.

So, this change fixes the SROC licence and charge version by adding the missing data to our testing set.