Finish up data for second scenario
We realised that we were missunderstanding licence references in the fixtures. The previous team were referring to things created in their test permits fixture and things in test licences as 'licences'.

So, when we tried to get CRMv1 fixtures working we couldn't understand why our `srocLicence02` wasn't being recognised. So as part of this change we've renamed the references in sroc-permits to `$srocPermit`.

Other than that, we completed the charge versions for scenario 2 (presroc and sroc). We also updated some values to make it clearer what scenario the data is linked to.