Add new presroc billing acc. change to fixture (#1870)
https://eaflood.atlassian.net/browse/WATER-3787

We're adding to our new SROC fixture set a pres-roc licence that we will then update the billing account on.

In the UI this would mean updating both pre-sroc and sroc charge versions. In this scenario, the 'licence' would be picked up for SROC supplementary billing because we added a new SROC charge version. But the new presroc charge version would be ignored.

** Notes

We updated a number of the references and things like `description` to make what scenario they relate to more meaningful.

We also fixed some issues with our existing SROC licence and charge version. We spotted that the Ui was not showing the charge element section for the SROC charge version we created.

We realised that SROC charge versions must have a charge_purpose created as well. And to create the charge_prupose we would need some purposes.

So, this change also fixes the SROC licence and charge version by adding the missing data to our testing set.