Add SROC sup. billing scenario 3 fixture data
https://eaflood.atlassian.net/browse/WATER-3787

This addes the fixture data for our 3rd scenario; a licence with both a presroc and sroc charge version. Only in our test for this the presroc charge version is beyond the 5 year window. So, when the change is made it should _not_ be included in a presroc bill run.

The sroc charge version should still be though!