Add SROC sup. billing scenario 3 fixture data (#1871)
https://eaflood.atlassian.net/browse/WATER-3787

This adds the fixture data for our 3rd scenario; a licence with a PRESROC and SROC charge version. Only in our test for this, the PRESROC charge version is beyond the 5-year window. So, when the change is made it should _not_ be included in a PRESROC bill run.

The SROC charge version should still be though!