Data migration fix for leap year charge purposes (#1894)
https://eaflood.atlassian.net/browse/WATER-3827

> Users can search for and find `7/35/01/*S/001`2 without issue and see the charge versions screen. But when the select 'View' in the first charge information record they get a "Sorry, the requested URL was rejected error".

We've investigated the issue and identified the cause. The charge version has been set up with 2 `charge_purposes`:

- Spray irrigation at Marsh Dykes at Frostenden - 1-Apr to 30-Sep
- Spray irrigation storage at Marsh Dykes at Frostenden - 1-Nov to 29-Feb

When the UI attempts to show this information it makes a 'date' out of the abstraction period by bolting the current year onto the `start` and `end` days and months.

In this case, it means it's trying to make a 'date' of **29-Feb-2022**, which doesn't exist hence things blow up. This is another example of an unhandled exception leaking through so users get the SilverLine WAF error page rather than our standard one. We found 3 other licences affected by this.

When we played around with it if we changed the abstraction period to 1-Nov to 28-Feb the UI no longer errored. We've discussed this with the users and they agree, an abstraction end date of **29-Feb** is a data error and should be fixed. The error only seems to apply to SROC charge versions so we think these records got in through the charge upload process.

As the business team is now finished with it, we don't expect any more problem records. Hence this change just creates a data migration to fix the problem records.