Bill numbers not showing on Bills Screen (#1908)
https://eaflood.atlassian.net/browse/WATER-3848

This could just be a timing issue as all the other Bill Runs and regions are showing OK, but currently for the Anglian Region the Bill number is not showing in the Bills screen on the Licences for the Anglian Annual Bill Run.

When pushing the fix to GitHub the CI was failing with the following error: `ERROR: stdout maxBuffer length exceeded Error: Process completed with exit code 1.`. This was being caused by a) the number of migrations which now exist in this repo and b) that the `-verbose` flag was set in `migrate.js`. We removed the verbose flag from the code and have created issue https://github.com/DEFRA/water-abstraction-team/issues/65 to do the same in the other repos