Revised approach to `prepare-transactions`
We realised that we didn't need to pass the batch type or scheme to the `prepare-transactions` job as it loaded the batch as part of its handler, and we can therefore add that to its `returnValue`. We therefore rework the logic accordingly.

As part of this we noticed that the tests did not correctly reflect the behaviour of the code -- the tests only had an experiment covering "when there are no transactions in the batch" when in fact the code has separate checks for the number of billing transaction ids (returned from the handler) and the number of transactions in the batch (freshly read in `onComplete`). We therefore rearrange the tests to match this.