Add new SROC test fixture (#1924)
https://eaflood.atlassian.net/browse/WATER-3834

We need to create a scenario where we have a licence with a charge version that has multiple charge categories (elements). Then linked to those categories are multiple charge purposes with different abstraction dates.

The intent is to help us confirm we are calculating the abstraction period for each category (element) in a charge version correctly. In this change, we are currently able to generate a new SROC licence with multiple categories, and one of those categories has multiple purposes (the other just has one).

It's a start but doesn't actually represent a real-life scenario. This is because the category is not defined through fixtures but via the logic in the fixture loader. That's going to take a bit more unpicking so we'll tackle that in another PR and update this fixture accordingly.

The key thing is this gets us started (and doesn't break when loaded!) as the structure is correct.