Implement solution
We're having to go with just increasing the number of charge categories available for selection from the table instead of making it truly dynamic. Re-reading our notes we'd forgotten what a problem it was just getting charge categories to work in the first place!

So, we've gone with 3 which can be picked from in the fixtures. It will now be easy enough to expand this in the future if needed.