Make more charge categories available in fixtures (#1944)
https://eaflood.atlassian.net/browse/WATER-3834

When working on the various scenarios needed to calculate the abstraction period for a charge version in a given billing period we realised there is one we could not cover.

A `charge_version` can be linked to multiple `charge_elements` and each charge element can be linked to different `billing_charge_categories`. But the current fixture logic only makes one  `billing_charge_category` available to use.

This change makes it possible to pick from 3 different `billing_charge_categories` (4.1.1, 4.2.1 and 4.2.10) in our fixtures.