Add new queued batch status (#1965)
https://eaflood.atlassian.net/browse/WATER-3877

When a supplementary bill run is started, we'll be creating both a PRESROC and SROC one at the same time.

The PRESROC one will then begin `processing` (labelled as BUILDING in the UI) whilst the SROC one is `queued` waiting to be processed.

At the moment though, all we can do is initialise the SROC bill run with a status of processing as that is all that is available to us. In the UI this will give the impression both are building when their not.

So, we need to add a new `queued` status to the code and DB. We'll make it so all new bill runs are given this status initially. Then, once processing starts the status will be updated. This is what happens at the moment. The bill run is created and then a job is added to a 'queue' for it to be processed.