Removing Lodash 'Pick' function (#1962)
https://github.com/DEFRA/water-abstraction-team/issues/25

Lodash is used a lot in our codebase. Although we are not totally removing this module, some of the functions in the code using lodash are unnecessary. Here we are removing the unnecessary ones are replacing it with vanilla js.# Please enter the commit message for your changes. Lines starting