Add timeout
Normally, we make this a configurable option, mainly for the purposes of testing that our requests behave in the way we expect.

However, the current tests are limited to checking whether we apply the proxy agent settings correctly or not. There is nothing that focuses on behaviour and little value in expanding the effort to add those kinds of tests to the legacy code base now.